requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.1
cuda-toolkit==13.0.2
cycler==0.12.1
decorator==4.4.2
edge-tts==7.2.8
exceptiongroup==1.3.1
facenet-pytorch==2.6.0
fastapi==0.135.3
fer==25.10.3
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.33
ml_dtypes==0.5.4
more-itertools==11.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.1
namex==0.1.0
networkx==3.4.2
numba==0.65.0
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==10.2.0
polars==1.39.3
polars-runtime-32==1.39.3
proglog==0.1.12
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.24
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scipy==1.15.3
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
sounddevice==0.5.5
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
tensorflow==2.21.0
termcolor==3.3.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0+cpu
torchaudio==2.11.0+cpu
torchvision==0.26.0+cpu
tqdm==4.67.3
transformers==5.5.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.43.0
webrtcvad==2.0.10
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
